requirements.txt
flask
pandas
numpy
scikit-learn
torch
networkx
pyvis
pytesseract
pillow
matplotlib
seaborn
transformers